% Auto-generated: do not edit by hand
\name{locCarousel}

\alias{locCarousel}

\title{Carousel component}

\description{
Carosel is used to make pretty carousels Need to include some extra css when using this
}

\usage{
locCarousel(children=NULL, id=NULL, arrows=NULL, autoplay=NULL,
center_mode=NULL, center_padding=NULL, class_name=NULL,
dots=NULL, infinite=NULL, responsive=NULL,
slides_to_scroll=NULL, slides_to_show=NULL, speed=NULL,
style=NULL, swipe_to_slide=NULL, variable_width=NULL,
vertical=NULL)
}

\arguments{
\item{children}{Your carousel is vertical}

\item{id}{Id of the element}

\item{arrows}{Arrows to control carousel}

\item{autoplay}{If the carousel will start to play automatically}

\item{center_mode}{To centralize the carousel}

\item{center_padding}{Padding on the sides}

\item{class_name}{Style class of the component.}

\item{dots}{Dots under carousel}

\item{infinite}{If the carousel content will repeat in a loop}

\item{responsive}{Settings of breakpoints}

\item{slides_to_scroll}{How many slides will scroll when you swipe the carousel}

\item{slides_to_show}{How many slides you want to show}

\item{speed}{Speed of autoplay}

\item{style}{Inline style of the component.}

\item{swipe_to_slide}{If you can slide to scroll the slides}

\item{variable_width}{The slides width varies according to the screen size}

\item{vertical}{If your carousel is vertical}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

