% Auto-generated: do not edit by hand
\name{locStudentOverviewCardNoBars}

\alias{locStudentOverviewCardNoBars}

\title{StudentOverviewCardNoBars component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
locStudentOverviewCardNoBars(id=NULL, border=NULL, class_name=NULL, data=NULL, name=NULL,
shown=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{border}{Character. Border of the card.}

\item{class_name}{Character. Classes for the outer most card.}

\item{data}{Named list. The data displayed on the card.}

\item{name}{Character. A name that will be printed when this component is rendered.}

\item{shown}{Unnamed list. The shown items displayed on the card.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

